fix(gate): split DB-fallback test to new file, add agent reports, fix stale comment (#612)
- Move test_handle_path_db_fallback_cold_start_finds_path to new
  graph_read_path_db_tests.rs (graph_read_path_tests.rs was 655 lines,
  now 497 — under 500-line limit)
- Add product/features/bugfix-612/ with investigator and rust-dev reports
  containing required Knowledge Stewardship sections
- Update stale AC-31 comment in test_tools.py (xfail reference removed)
